requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.18.11
botocore==1.21.65
CacheControl==0.13.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.0
daphne==4.0.0
Django==4.2.1
django-cors-headers==4.1.0
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
firebase-admin==6.1.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
googlemaps==4.10.0
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==0.10.0
jws==0.1.3
msgpack==1.0.5
oauth2client==4.1.3
proto-plus==1.22.2
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
Pyrebase4==4.7.0
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2023.3
razorpay==1.3.0
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
zope.interface==6.0
python-dotenv